Refactor experimental framework into a package
