--- experiment/__init__.py
@@ -1,4 +1,0 @@
-from config import ClassifierConfig, DatasetConfig, ExperimentConfig
-from runner import ExperimentManager
-
-__all__ = ["ClassifierConfig", "DatasetConfig", "ExperimentConfig", "ExperimentManager"]
