--- experiment/schema.json
@@ -1,1 +1,0 @@
-{"$schema":"http://json-schema.org/draft-07/schema#","title":"Gender Bias Experiment Results","description":"Results from experiments analyzing gender bias in facial classification models under various conditions","type":"array","items":{"type":"object","required":["metadata","bias_analyses"],"properties":{"metadata":{"type":"object","description":"Information about the experimental conditions","required":["id","male_ratio","female_ratio","masked_feature","masked_gender"],"properties":{"id":{"type":"string","description":"Unique MD5 hash identifier for the experiment","pattern":"^[0-9a-f]{32}$"},"male_ratio":{"type":"integer","description":"Percentage of male samples in the training dataset (0-100)","minimum":0,"maximum":100},"female_ratio":{"type":"integer","description":"Percentage of female samples in the training dataset (0-100)","minimum":0,"maximum":100},"masked_feature":{"type":["string","null"],"description":"Facial feature that was masked during training, or null if no masking","enum":[null,"left_eye","right_eye","nose","lips","left_cheek","right_cheek","chin","forehead","left_eyebrow","right_eyebrow"]},"masked_gender":{"type":["integer","null"],"description":"Gender for which the feature was masked (0 for male, 1 for female), or null if no masking","enum":[null,0,1]}}},"bias_analyses":{"type":"array","description":"Results from multiple experiment replicates","items":{"type":"object","required":["disparity_scores","feature_analyses"],"properties":{"disparity_scores":{"type":"object","description":"Overall bias measurement scores","required":["biasx","equalized_odds"],"properties":{"biasx":{"type":"number","description":"BiasX score - average of individual feature bias scores (0-1 scale, higher means more bias)","minimum":0,"maximum":1},"equalized_odds":{"type":"number","description":"Maximum difference in true/false positive rates between genders (0-1 scale, higher means more bias)","minimum":0,"maximum":1}}},"feature_analyses":{"type":"object","description":"Bias measurements for individual facial features","patternProperties":{"^(left_eye|right_eye|nose|lips|left_cheek|right_cheek|chin|forehead|left_eyebrow|right_eyebrow)$":{"type":"object","required":["bias_score","male_probability","female_probability"],"properties":{"bias_score":{"type":"number","description":"Absolute difference in activation probabilities between genders for this feature (0-1 scale)","minimum":0,"maximum":1},"male_probability":{"type":"number","description":"Probability this feature appears in activation maps for misclassified male images","minimum":0,"maximum":1},"female_probability":{"type":"number","description":"Probability this feature appears in activation maps for misclassified female images","minimum":0,"maximum":1}}}}}}},"minItems":1}}}}
