+++ experiment/utils.py
@@ -1,0 +1,34 @@
+import logging
+import os
+from datetime import datetime
+
+
+def setup_logger(name="biasx_framework", log_dir="logs", console_level=logging.INFO, file_level=logging.DEBUG):
+    """Configure and return a logger that writes to both console and a timestamped log file."""
+    os.makedirs(log_dir, exist_ok=True)
+
+    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
+    log_filename = os.path.join(log_dir, f"{name}_{timestamp}.log")
+
+    logger = logging.getLogger(name)
+    logger.setLevel(min(console_level, file_level))
+
+    if logger.hasHandlers():
+        logger.handlers.clear()
+
+    file_handler = logging.FileHandler(log_filename)
+    file_handler.setLevel(file_level)
+
+    console_handler = logging.StreamHandler()
+    console_handler.setLevel(console_level)
+
+    formatter = logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(funcName)s:%(lineno)d - %(message)s")
+    file_handler.setFormatter(formatter)
+    console_handler.setFormatter(formatter)
+
+    logger.addHandler(file_handler)
+    logger.addHandler(console_handler)
+
+    logger.info(f"Logging initialized: file output to {log_filename}")
+
+    return logger
